% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquaculture.R
\name{aquaculture_licenses}
\alias{aquaculture_licenses}
\title{Download BC Aquaculture license holders data}
\usage{
aquaculture_licenses()
}
\description{
This function enables download of BC Aquaculture
license holders from Fisheries and Oceans Canada.
Source metadata: https://open.canada.ca/data/en/dataset/522d1b67-30d8-4a34-9b62-5da99b1035e6#wb-auto-4
}
\keyword{aquaculture}
