% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfor.R
\docType{package}
\name{dfor-package}
\alias{dfor}
\alias{dfor-package}
\title{dfor: Department of Fisheries and Oceans open data for R}
\description{
An R package for accessing open data from the Department
    of Fisheries and Oceans in Canada (Fisheries and Oceans Canada).
}
\details{
The only R package for accessing open data from Fisheries and Oceans Canada.
}
\keyword{DFO}
\keyword{canada}
\keyword{data}
\keyword{fisheries}
\keyword{oceans}
\keyword{open}
